Updated shipment entities. Added new migration called "Initial Shipment entities". Added two BaseEntities to reflects short and long id's and update other entities to use those, instead of current BaseEntity. Change Dto 's to reflect above Id type.

+++ PI.Contract/DTOs/AccountSettings/TimeZoneDto.cs
@@ -8,7 +8,7 @@ namespace PI.Contract.DTOs.AccountSettings
 {
      public class TimeZoneDto
     {
-        public long Id { get; set; }
+        public short Id { get; set; }
         public string TimeZoneCode { get; set; }
         public string CountryName { get; set; }
     }
